requirements.txt
flask>=3.0.0
openpyxl>=3.1.2
pandas>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=10.2.0
werkzeug>=3.0.1
gunicorn>=21.2.0

